Fix formatting on some things
